Mostly in sync with CJLST's 2P2F CR.
3P1F was surprisingly constant. Will implement viable 4-lep combos of 3P1F tomorrow.
