+++ constants/particleprops.py
@@ -1,0 +1,2 @@
+# From the PDG: m(Z) = 91.1876 +- 0.0021 GeV
+ZMASS_PDG = 91.1876  # GeV.
